api/requirements.txt
python-dotenv
Flask
oauthlib
Flask-Login
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
flask-session
cryptography
Flask-Cors
pytest
opencv-python
networkx
openai
polyline